requirements.txt
letta-client
python-decouple
discord
coloredlogs
peewee